src/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Alex Lutterloah <lutterloaha@vcu.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
matplotlib = "^3.8.2"
plotly = "^5.18.0"
jupyter = "^1.0.0"
tabulate = "^0.9.0"
markdown = "^3.5.1"
panda = "^0.3.1"
sqlalchemy = "^2.0.23"
pymysql = "^1.1.0"
mysql-connector-python = "^8.2.0"
python-dotenv = "^1.0.0"
pandas = "^2.1.3"
flask = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
